tag delete

+++ app/Http/Controllers/Admin/TagController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers\Admin;
 
 
+use App\Models\Article;
 use App\Models\Tag;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class TagController extends AdminBaseController
 {
@@ -19,6 +21,10 @@ class TagController extends AdminBaseController
 
     public function delete($id): JsonResponse
     {
+        $hasArticle = DB::table('article_tags')->where('tag_id',$id)->value('article_id');
+        if ($hasArticle) {
+            return response()->json(['msg'=>'该标签下还有文章'],403);
+        }
 
         Tag::where('id',$id)->delete();
         return response()->json();
